Use Keystone v3 API for user creation
This includes requiring a domain when creating a user. This will allow us to
control where users are created in a later patch.

Adding the token to the user creation call is required because of a bad interaction between OpenStackClient, os-client-config and keystoneclient when dealing with v2 authentication but v3 API calls. It will be cleaned up when we switch to v3 credentials.